refactor: restructure folders to isolate behaviors
