+++ package.json
@@ -2,7 +2,7 @@
   "name": "old-man-bot",
   "version": "0.0.2",
   "description": "I'm just an old bot, afraid of change.",
-  "main": "src/bot.ts",
+  "main": "src/OldManBot/OldMan.ts",
   "author": "cargillj",
   "watch": {
     "start": {
@@ -15,7 +15,7 @@
     "pre-commit-message": "echo Performing pre-commit checks... && exit 0",
     "build": "tsc -p .",
     "clean": "rimraf build",
-    "start": "NODE_ENV=PROD rimraf build && tsc -p . && node build/bot.js",
+    "start": "NODE_ENV=PROD rimraf build && tsc -p . && node build/OldManBot/OldMan.js",
     "test": "NODE_ENV=TEST jest",
     "watch": "npm-watch",
     "release": "standard-version"
