+++ src/DiscordBot/Managers/ResponseManager/Response.ts
