+++ src/DiscordBot/Managers/ResponseManager/index.ts
