+++ src/DiscordBot/index.ts
@@ -1,0 +1,3 @@
+import DiscordBot from './DiscordBot'
+
+export default DiscordBot
