--- src/tests/responses/offensives.test.ts
@@ -1,7 +1,0 @@
-import { boomer } from '../../responses/offensives'
-
-describe('offensives', () => {
-  test('ok boomer', () => {
-    expect(boomer.isTriggered('ok boomer')).toBeTruthy()
-  })
-})
